+++ sdk/metric/x/doc.go
@@ -1,0 +1,8 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package x contains experimental metric SDK options and features.
+//
+// These features are experimental and under development. They may be changed
+// in backwards-incompatible ways or removed in future releases.
+package x